Logo
Many of the most impactful and successful logos in history are surprisingly simple. From Nike’s single swoosh to Apple’s eponymous design, simple logos are easy to recognize and remember. Simplicity is a key ingredient for logos because most consumers only focus on a logo for a short time. A simple design can express your brand’s personality concisely and effectively.

Simple logos focus on highlighting the most important parts of a brand’s personality with limited real estate. This includes focusing more on aspects like colors and fonts as well as on distilling ideas into their simplest form. For instance, a symbol is a powerful way of creating simplicity, as it can instill a mental association with a particular set of values or ideas.

Other great designs for a simple logo include letter and word marks, which dispense with images and focus instead on communicating brand personality directly with fonts and colors. Overall, the most important aspect is to focus on using as few elements as possible to communicate your brand’s identity.