BuffKings logo design
The first quality great logos share is that they’re relevant to the markets their companies target. More importantly, they clearly communicate a brand’s personality and identity. A primary component is the use of colors in your logo, which can trigger different emotions and show your brand’s personality to consumers. A company that sells toys for children may choose bright colors that communicate energy, fun, and excitement.

The second important component is the font used in the logo or word mark. Fonts help communicate your brand’s tone and its values, which ultimately help define your personality better. More angular and thin fonts are ideal for highlighting a company that works in technology while softer cursives are excellent for companies that work in jewelry or women’s products.

Finally, choosing the right symbol is a key aspect of establishing a visual anchor for your logo. Symbols are important aspects of a logo because they can be used by themselves as a simpler version of your logo. Symbols are also important when building connections between your brand and the ideas and values behind it.